+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Improve choose_variables to control the order of the returned variables
     - The return type is now a list (and not a Tuple)
     - The order of the returned list is based on the variables argument. If it is not supplied, it is based on column names in the DataFrames. The df_for_var_order arg controls which df to use.
-
+- The `_prepare_input_model_matrix` and downstream functions (e.g.: `model_matrix`, `sample.outcomes().mean()`, etc) can now handle DataFrame with special characters in the column names, by replacing special caracters with '_' (or '_i', if we end up with columns with duplicate names). It also handles cases in which the column names have duplicates (using the new `_make_df_column_names_unique` function).
 
 0.5.0 (2023-03-06)
 ==================